working pure-Go implementation of range lease storage (toggled by build tags)
