--- plugins/range/boltdb_test.go
@@ -1,7 +1,0 @@
-// Copyright 2018-present the CoreDHCP Authors. All rights reserved
-// This source code is licensed under the MIT license found in the
-// LICENSE file in the root directory of this source tree.
-
-//go:build !cgo
-
-package rangeplugin
